[SUPER] 관리자 삭제 기능 구현 - JWT Token subject에 email이 아닌 adminId가 들어가도록 수정

+++ src/main/java/com/highthon/highthon3server/controller/AdminController.java
@@ -1,9 +1,13 @@
 package com.highthon.highthon3server.controller;
 
-import com.highthon.highthon3server.dto.auth.AdminSignupDto;
+import com.highthon.highthon3server.exception.SelfHandleException;
+import com.highthon.highthon3server.security.JwtTokenUtil;
 import com.highthon.highthon3server.service.admin.AdminService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
 
 @RestController
 public class AdminController {
@@ -11,8 +15,17 @@ public class AdminController {
     @Autowired
     private AdminService adminService;
 
+    @Value("${jwt.tokenHeader}")
+    private String tokenHeader;
+
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
+
     @DeleteMapping("/admin/{adminId}")
-    public void deleteAdmin(@PathVariable("adminId") String adminId) {
+    public void deleteAdmin(@PathVariable("adminId") String adminId, HttpServletRequest request) {
+        String token = request.getHeader(tokenHeader).substring(7);
+        String myId = jwtTokenUtil.getAdminIdFromToken(token);
+        if (myId.equals(adminId)) throw new SelfHandleException();
         adminService.deleteAdmin(adminId);
     }
 }
